Update to speed up
Now the proxy speed is at the same level with the one w/o proxy. Both at 3247K.

In this design, we actively try to parse the header of the response when it is available. And if we get the content-length, we can use it to early determine the transfer is over.